refactor: rename experimental _feature.py to _calculate_image_features.py
Name the module after its public function, consistent with the sibling
_detect_tissue.py / _make_tiles.py / _qc_image.py modules in experimental/im.
