+++ src/squidpy/experimental/im/__init__.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
+from ._calculate_image_features import calculate_image_features
 from ._detect_tissue import (
     BackgroundDetectionParams,
     FelzenszwalbParams,
     WekaParams,
     detect_tissue,
 )
-from ._feature import calculate_image_features
 from ._make_tiles import make_tiles, make_tiles_from_spots
 from ._qc_image import qc_image
 from ._qc_metrics import QCMetric
